1920 - Fix stale Float32 fixture in adjust_managerial_roles test
registered_manager_count was still typed Float32 in this one fixture, unlike
its sibling calculate_reg_man_difference_schema (already UInt32, matching the
real upstream dtype). adjust_managerial_roles's test is the only one that
runs the full calculate_reg_man_difference -> calculate_non_rm_managerial_distribution
-> distribute_rm_difference chain together, so it never exercised the
UInt32-promotion scenario this ticket's dtype fix addresses.

+++ projects/_03_independent_cqc/unittest_data/polars_ind_cqc_test_file_schemas.py
@@ -1486,7 +1486,7 @@ class EstimateFilledPostsByJobRoleEstimateUtilsSchemas:
                 CatVals.main_job_role_labels_column_values.categorical_values
             ),
             IndCQC.estimate_filled_posts_by_job_role: pl.Float32,
-            IndCQC.registered_manager_count: pl.Float32,
+            IndCQC.registered_manager_count: pl.UInt32,
         }
     )
     expected_adjust_managerial_roles_schema = pl.Schema(
